requirements.txt
# Core Dashboard Framework
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Statistical Analysis & A/B Testing
scipy==1.11.4
statsmodels==0.14.1

# Environment Variables
python-dotenv==1.0.0

# Utilities
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
